requirements.txt
langchain==0.0.280
pinecone-client==2.2.3
streamlit==1.30.0 
unstructured==0.6.0
unstructured[pdf]==0.6.0
openai==0.27.0
chromadb==0.3.26
tiktoken==0.3.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==0.8.11
validators==0.20.0
